COMMERCE-1499 talend - component - runtime - TEST - move LiferaySink implementation to runtime package
